TEST: Split pytests into separate files
