+++ test/test_0000_imports.py
@@ -1,0 +1,35 @@
+# ==========================================================================
+#
+#   Copyright NumFOCUS
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#          https://www.apache.org/licenses/LICENSE-2.0.txt
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ==========================================================================
+
+
+def test_0000_imports():
+    """Purpose: Test that needed packages are available"""
+    import al_bench as alb
+    import h5py as h5
+    import numpy as np
+    import os
+    import random
+    import re
+    import tensorflow as tf
+    import torch
+
+    pass
+
+
+if __name__ == "__main__":
+    test_0000_imports()
